ref(types): Move sampling types to dedicated module
Move `SampleRand` and `InvalidSampleRandError` from `protocol/v7.rs` into a new private top-level `sampling` module, re-exported from the old `protocol::v7` path, so the public API is unchanged. This is primarily code motion without any behavioral changes; the only non-mechanical edits are the imports the new module needs (`serde`, `thiserror`, and some `std::fmt`/`std::str` items).

This prepares the module for upcoming sampling work related to propagating the sampling seed (`sample_rand`) with traces.

References [#735](https://github.com/getsentry/sentry-rust/issues/735)
References [RUST-1](https://linear.app/getsentry/issue/RUST-1)
